Allow users to disable default page tracking
We wanted to continue using the defaultPageTrack name but default to
enabled if this was not included in the segment configuration. To know
if the user explicitly disabled the feature, we check for the false, not
falsey, value of `config.segment.defaultPageTrack`.

We felt that abstracting this logic into two function
(pageTrackEnabled/pageTrackDisabled) helped readability of the code.